requirements.txt
Django==4.2.29
django-taggit==4.0.0
django-extensions==3.2.3
django-bootstrap3==23.1
python-debian==1.0.1
defusedxml==0.7.1
PyYAML==6.0.2
requests==2.32.4
colorama==0.4.6
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
django-filter==25.1
humanize==4.12.1
version-utils==0.3.2
python-magic==0.4.27
gitpython==3.1.44
tenacity==8.2.3
celery==5.4.0
redis==6.4.0
django-celery-beat==2.7.0
tqdm==4.67.1
cvss==3.4
zstandard==0.25.0
django-tables2==2.8.0
django-select2==8.4.1
